+++ gatsby-node.js
@@ -153,6 +153,8 @@ exports.createPages = async ({ graphql, actions }) => {
     }
   `);
 
+  // For every node in "area-indices" create a parentId to child Ids look up
+  // data structure
   const childAreas = {};
   result.data.allMdx.edges.map(({node})=>{
     const parentId = node.fields.parentId;
@@ -169,18 +171,18 @@ exports.createPages = async ({ graphql, actions }) => {
   // Create each index page for each leaf area
   const { createPage } = actions;
   for (const {node} of result.data.allMdx.edges) {
-    const childAreaPaths = childAreas[node.fields.pathId] ? childAreas[node.fields.pathId] : [];
+    // For a given area indices, list out all the possible path strings for
+    // the children areas
+    const childAreaPathIds = childAreas[node.fields.pathId] ? childAreas[node.fields.pathId] : [];
     createPage({
       path: node.fields.slug,
       component: path.resolve(`./src/templates/leaf-area-page-md.js`),
       context: {
         legacy_id: node.frontmatter.metadata.legacy_id,
-        // climbs: node.climbs,
-        // name: node.area_name,
         slug: node.fields.slug,
         pathId: node.fields.pathId,
         possibleParentPaths: node.fields.possibleParentPaths,
-        childAreaPaths: childAreaPaths
+        childAreaPathIds: childAreaPathIds
       },
     });
   }
