+++ src/js/utils.js
@@ -117,3 +117,34 @@ export const computeClimbingPercentsAndColors = (climbs) => {
     colors
   };
 };
+
+export const computeStatsBarPercentPerAreaFromClimbs = (climbs) => {
+
+  const areasToClimbs = {};
+  const areasToPercentAndColors = {};
+
+  // map each climb to the area 
+  climbs.edges.map(({node})=>{
+    const parentId = node.fields.parentId;
+    if (areasToClimbs[parentId]) {
+      areasToClimbs[parentId].push(node.frontmatter)
+    } else {
+      areasToClimbs[parentId] = [node.frontmatter];
+    }
+  });
+
+  // compute the stats and percent per area
+  // do a little formatting to  reuse the helper function
+  Object.keys(areasToClimbs).map((key)=>{
+    const formatted = areasToClimbs[key].map((c)=>{
+      return {
+        node: {
+          frontmatter: c
+        }
+      };
+    });
+    areasToPercentAndColors[key] = computeClimbingPercentsAndColors(formatted);
+  })
+
+  return areasToPercentAndColors;
+};
